+++ src/context.js
@@ -9,6 +9,7 @@ export const AppContext = createContext({
 
     vtcLogo: localStorage.getItem("cache-logo"), setVtcLogo: () => { },
     vtcBanner: localStorage.getItem("cache-banner"), setVtcBanner: () => { },
+    vtcBackground: localStorage.getItem("cache-background"), setVtcBackground: () => { },
 
     apiConfig: null, setApiConfig: () => { },
     webConfig: null, setWebConfig: () => { },
@@ -52,6 +53,7 @@ export const AppContextProvider = ({ children }) => {
 
     const [vtcLogo, setVtcLogo] = useState(localStorage.getItem("cache-logo"));
     const [vtcBanner, setVtcBanner] = useState(localStorage.getItem("cache-banner"));
+    const [vtcBackground, setVtcBackground] = useState(localStorage.getItem("cache-background"));
 
     const [apiConfig, setApiConfig] = useState(null);
     const [webConfig, setWebConfig] = useState(null);
@@ -242,6 +244,7 @@ export const AppContextProvider = ({ children }) => {
 
         vtcLogo, setVtcLogo,
         vtcBanner, setVtcBanner,
+        vtcBackground, setVtcBackground,
 
         apiConfig, setApiConfig,
         webConfig, setWebConfig,
@@ -267,7 +270,7 @@ export const AppContextProvider = ({ children }) => {
         dlogDetailsCache, setDlogDetailsCache, loadDlogDetails,
         economyCache, setEconomyCache,
         allUsersCache, setAllUsersCache, loadAllUsers
-    }), [apiPath, apiVersion, apiConfig, vtcLogo, vtcBanner, webConfig, languages, allRoles, allPerms, allRanks, users, userProfiles, memberUIDs, curUID, curUser, curUserPerm, curUserBanner, userSettings, announcementTypes, applicationTypes, divisions, dlogDetailsCache, economyCache, allUsersCache]);
+    }), [apiPath, apiVersion, apiConfig, vtcLogo, vtcBanner, vtcBackground, webConfig, languages, allRoles, allPerms, allRanks, users, userProfiles, memberUIDs, curUID, curUser, curUserPerm, curUserBanner, userSettings, announcementTypes, applicationTypes, divisions, dlogDetailsCache, economyCache, allUsersCache]);
 
     return (
         <AppContext.Provider value={value}>