+++ src/designs.js
@@ -78,7 +78,7 @@ function intToHex(intValue) {
     return hexValue;
 }
 
-export function getDesignTokens({ webConfig }, { themeSettings, setThemeSettings }, customMode, mode, use_custom_theme = false, theme_background = null, theme_main = null, darken_ratio = null, font_size = "regular") {
+export function getDesignTokens({ vtcBackground, webConfig }, { themeSettings, setThemeSettings }, customMode, mode, use_custom_theme = false, theme_background = null, theme_main = null, darken_ratio = null, font_size = "regular") {
     if (use_custom_theme === true) {
         if (vars.userLevel < 2)
             use_custom_theme = false;
@@ -92,7 +92,7 @@ export function getDesignTokens({ webConfig }, { themeSettings, setThemeSettings
         theme_main = webConfig.theme_main_color;
         darken_ratio = webConfig.theme_darken_ratio;
     }
-    if (use_custom_theme === "vtcbg" && vars.vtcLevel >= 1 && vars.dhvtcbg !== "") {
+    if (use_custom_theme === "vtcbg" && vars.vtcLevel >= 1 && vtcBackground !== "") {
         if (darken_ratio === null) darken_ratio = 0.4;
         if (mode === "dark") {
             if (theme_background === null) theme_background = "#212529";
@@ -103,8 +103,8 @@ export function getDesignTokens({ webConfig }, { themeSettings, setThemeSettings
         }
         theme_background = theme_background.substring(0, 7) + intToHex(darken_ratio * 100);
         theme_main = theme_main.substring(0, 7) + intToHex(darken_ratio * 100);
-        if (setThemeSettings !== undefined && themeSettings.bg_image !== vars.dhvtcbg) { // ensure called from <App>
-            setThemeSettings(prev_settings => ({ ...prev_settings, bg_image: vars.dhvtcbg }));
+        if (setThemeSettings !== undefined && themeSettings.bg_image !== vtcBackground) { // ensure called from <App>
+            setThemeSettings(prev_settings => ({ ...prev_settings, bg_image: vtcBackground }));
             return; // we know there'll be a re-render
         }
     }